initial stab at organising the code in a nicer way so others can navigate it
