+++ elementsToString.js
@@ -1,0 +1,5 @@
+module.exports = function elementsToString(els) {
+  return els.toArray().map(function (el) {
+    return el.outerHTML.replace(el.innerHTML, '');
+  }).join(', ');
+}
