+++ jquery.js
@@ -1,0 +1,4 @@
+module.exports = 
+  typeof $ === 'undefined'
+    ? require("jquery")
+    : window.$;
